Minor fix. Moved try_zome_call() to setup.rs
